+++ release_notes.md
@@ -1,10 +1,38 @@
 Template
 =============================================================
 x.x.x Release notes (yyyy—MM-dd)
+================================
 
 C++ (core)
 -----------
 The C++ API has been updated and your code will break!
+
+### Bugfixes:
+
+* None.
+
+### API breaking changes:
+
+* `???`
+
+### Enhancements:
+
+* `???`
+
+-----------
+
+### Internals:
+
+* `???`
+
+----------------------------------------------
+
+x.x.x Release notes (yyyy—MM-dd)    <<<------------ PLEASE NOTE, THIS IS THE NEXT RELEASE!!!!!!
+================================
+
+C++ (core)
+-----------
+Adding Table::get_index_in_parent(). It was needed for implicit transactions.
 
 ### Bugfixes: